test(backend-metal): d_input 診断ベンチの型不一致とdoc記述を是正する
- mnist_scale_train_reuse_bench.rs の run_step_measure_backward
  クロージャの戻り値型を (f64, u64, u64, u64) から
  (f64, usize, usize, usize) へ修正する。BatchCountersSnapshot
  の差分式（usize 同士の減算）が u64 と型不一致になり
  aarch64-apple-darwin ターゲットで E0308 が発生していた
  （codex-review 指摘。cargo check --target aarch64-apple-darwin
  で再現・是正後の解消を確認済み）。
- resident_lhs_dinput_phase_bench.rs 冒頭 doc comment を修正し、
  dispatch_strided_bias_act_prepared の実際の処理
  （encode + ctx.synchronize() のみ・readback は含まない）に
  合わせる（gemm.rs:3146-3161 で確認済み）。

Refs #1562

+++ crates/facade/tests/mnist_scale_train_reuse_bench.rs
@@ -494,7 +494,7 @@ fn mnist_scale_train_reuse_metal_backward_dinput_phase() {
     // 返す。forward・update 自体は計測窓の外（`Instant` 計測は
     // backward_device_param_store 呼び出しの前後のみ）。
     let mut run_step_measure_backward =
-        |store: &mut fandhe_ai::DeviceParamStore| -> (f64, u64, u64, u64) {
+        |store: &mut fandhe_ai::DeviceParamStore| -> (f64, usize, usize, usize) {
             let tape = fandhe_ai::tape_for(device).unwrap();
             let x = tape.var(&x_data);
             let y = tape.var(&y_data);
